changes suggested by ramya
