+++ src/goTest.ts
@@ -10,9 +10,9 @@ import path = require('path');
 import vscode = require('vscode');
 import util = require('util');
 import os = require('os');
-import { getGoRuntimePath } from './goPath';
+import { parseEnvFile, getGoRuntimePath, resolvePath } from './goPath';
+import { getToolsEnvVars } from './util';
 import { GoDocumentSymbolProvider } from './goOutline';
-import { getTestEnvVars } from './util';
 
 let outputChannel = vscode.window.createOutputChannel('Go Tests');
 
@@ -137,6 +137,21 @@ export function testWorkspace(goConfig: vscode.WorkspaceConfiguration, args: any
 	}).then(null, err => {
 		console.error(err);
 	});
+}
+
+export function getTestEnvVars(): any {
+	const config = vscode.workspace.getConfiguration('go');
+	const toolsEnv = getToolsEnvVars();
+	const testEnv = config['testEnvVars'] || {};
+
+	let fileEnv = {};
+	let testEnvFile = config['testEnvFile'];
+	if (testEnvFile) {
+		testEnvFile = resolvePath(testEnvFile, vscode.workspace.rootPath);
+		fileEnv = parseEnvFile(testEnvFile);
+	}
+
+	return Object.assign({}, toolsEnv, fileEnv, testEnv);
 }
 
 /**
